benchmark working

+++ src/domain/bound.rs
@@ -3,10 +3,10 @@
 #[derive(Copy)]
 #[derive(PartialEq)]
 pub struct Bound {
-    pub x0: f32,
-    pub x1: f32,
-    pub y0: f32,
-    pub y1: f32,
-    pub z0: f32,
-    pub z1: f32
+    pub x0: i32,
+    pub x1: i32,
+    pub y0: i32,
+    pub y1: i32,
+    pub z0: i32,
+    pub z1: i32
 }
