+++ src/domain/entity.rs
@@ -4,7 +4,7 @@
 #[derive(PartialEq)]
 pub struct Entity {
     pub id: i32,
-    pub x_coordinate: f32,
-    pub y_coordinate: f32,
-    pub z_coordinate: f32
+    pub x_coordinate: i32,
+    pub y_coordinate: i32,
+    pub z_coordinate: i32
 }
